net6 - get rid of non-integer frame rates, ffMPEG does not support

+++ src-cs/CFForm/Document.cs
@@ -326,8 +326,8 @@ namespace CFForm
                     height = Properties.Settings.Default.RenderHeight,
                     borderSize = Properties.Settings.Default.BorderWidth,
                     minimumSize = Properties.Settings.Default.MinimumSize,
-                    animateFrameCount = (int)(Properties.Settings.Default.AnimateLength *
-                                              Properties.Settings.Default.AnimateFrameRate)
+                    animateFrameCount = Properties.Settings.Default.AnimateLength *
+                                        Properties.Settings.Default.AnimateFrameRate
                 };
             } 
             set
